feat(fr): fix 404 tests for form explainers, move 404 code to navigation tests
