+++ test/browser_testing/features/loan_estimate.feature
@@ -104,10 +104,10 @@ Examples:
   | 3           | 2        |
   | 2           | 1        |
 
-@check_urls @loan_estimate
+@404 @loan_estimate
 Scenario Outline: Testing availability of pages on Loan Estimate
   Given I navigate to the "<page_name>" page
-  Then All links are working
+  Then Links are working without 404 errors
 
 Examples:
   | page_name             |
