+++ test/browser_testing/features/pages/navigation.py
@@ -6,6 +6,10 @@ from selenium import webdriver
 
 from pages.base import Base
 
+import requests
+
+# ELEMENT FOR LINKS
+LINK_TAG = '#primary-content a'
 
 class Navigation(Base):
 
@@ -48,3 +52,24 @@ class Navigation(Base):
         script = "arguments[0].scrollIntoView(true);"
         self.driver.execute_script(script, element)
         element.click()
+
+    def check_link_status_code(self, link):
+        try:
+            r = requests.head(link)
+            return r.status_code > 199 and r.status_code < 400
+        except requests.ConnectionError:
+            return False
+
+    def check_links_for_404s(self, base_url):
+        results = []
+        link_elements = self.driver.find_elements_by_css_selector( LINK_TAG )
+        for elem in link_elements:
+            link = elem.get_attribute('href')
+            is_local = base_url.startswith('http://localhost') or base_url.startswith('http://127.0.0.1')
+            if link and not link.startswith('tel') and not self.check_link_status_code(link):
+                # because when running tests locally links like localhost/es will return 404
+                # and there are too many of them
+                if is_local and link.startswith(base_url):
+                    results.append(link)
+                    print link
+        return results
